close contact page after moving to new org

+++ app/controllers/organizational_roles_controller.rb
@@ -13,17 +13,19 @@ class OrganizationalRolesController < ApplicationController
     Organization.transaction do
       from_org = Organization.find(params[:from_id])
       to_org = Organization.find(params[:to_id])
-      people = Person.find(params[:ids])
+      unless from_org == to_org
+        people = Person.find(params[:ids])
     
-      # Remove them from the current org
-      ContactAssignment.where(person_id: params[:ids], organization_id: from_org.id).destroy_all
-      people.each do |person|
-        from_org.remove_contact(person)
-      end
+        # Remove them from the current org
+        ContactAssignment.where(person_id: params[:ids], organization_id: from_org.id).destroy_all
+        people.each do |person|
+          from_org.remove_contact(person)
+        end
     
-      # Add them to the new org
-      people.each do |person|
-        to_org.add_contact(person)
+        # Add them to the new org
+        people.each do |person|
+          to_org.add_contact(person)
+        end
       end
     end
     render nothing: true
